Replace the "level of authentication too low" error
This removes a field from the error response, but the only app which
consumes this error (finder-frontend) doesn't use the response body,
it just looks at the status code.  So this isn't really a breaking
change.
